refactor: Update VisitorsComponent to display "All visitors loaded" message
This commit updates the VisitorsComponent to display a message when all visitors have been loaded. It adds a new property `isAllDone` to track if all visitors have been loaded and updates the HTML template to conditionally display the message using the `app-empty-screen` component. The message includes an icon, title, and color to match the design. This improves the user experience by providing feedback when all visitors have been loaded.