chore: Update Discord links in README.md
This commit updates the Discord links in the README.md file to point to the new Discord server URL "https://discord.langx.io". This change ensures that users can easily join the LangX Discord community and access support.

Note: This commit message is based on the code changes and recent repository commits.