feat: Add close() method to CopilotInstructionsComponent
This commit adds a new method `close()` to the `CopilotInstructionsComponent` in order to handle the closing of the component. The `close()` method calls the `onFinish()` function, which is an input property of the component. This enhancement improves the functionality of the Copilot instructions component by providing a way to close it programmatically.