feat: Add deleteIdentity method to AuthService
This commit adds a new method called deleteIdentity to the AuthService class in auth.service.ts. The deleteIdentity method takes an instance of the Models.Identity class as a parameter and returns an Observable<any>. It calls the deleteIdentity method of the api.account object and passes the $id property of the identity object as an argument. This change enhances the user management capabilities by allowing the deletion of identities.