feat: Add last seen card to account settings page
This commit adds a new card to the account settings page that displays the last seen information of the user. The card includes the date and time of the user's last activity, as well as a note indicating the time elapsed since the last activity. This feature enhances the user experience by providing visibility into the user's activity status.