fix: #340 Update listRooms method to include blocked users  🚀 [Feature] : Query for users that are not blocked by the current user #340
This commit modifies the listRooms method in the RoomService to include support for filtering out rooms that contain blocked users. It iterates through the blockedUsers array of the currentUser and adds a query to exclude rooms that contain those users. This change improves the functionality of the chat application by preventing blocked users from appearing in the list of rooms.