refactor: Update listRooms method to exclude blocked users 🚀 [Feature] : Query for users that are not blocked by the current user #340
This commit modifies the listRooms method in the RoomService to exclude rooms that contain blocked users. It removes the condition that checks if the room is in the currentUser's archivedRooms array. This change improves the functionality of the chat application by preventing blocked users from appearing in the list of rooms.