refactor: Update Room component onPress event handling
This commit updates the onPress event handling in the Room component to handle different actions based on the index of the pressed item in the ContextMenu. If the index is 0, the reply message is set to the current message. If the index is 1, the "Edit" action is logged. If the index is 2, the "Copy" action is logged. This change improves the functionality of the Room component and provides better user interaction.

Related to recent user commits:
- refactor: Update Room component to use themed ContextMenu
- feat: Add ContextMenu to Room component for message options
- refactor: Update dependencies for react-native-context-menu-view to version 1.16.0