feat: Update Room component to use react-native-get-random-values
This commit adds the "react-native-get-random-values" package as a dependency and imports it in the Room component. This package provides a reliable way to generate random values, which is used in the component for certain functionalities. This change enhances the functionality and reliability of the Room component.

Related to recent user commits:
- refactor: Use Clipboard to copy message text in Room component
- refactor: Update npm dependencies for react-native-clipboard to version 1.14.1