fix: web version of contextmenu, Update Room component to use Ionicons for message options
This commit updates the Room component to use the Ionicons package from "@expo/vector-icons" for the message options in the ContextMenu. It replaces the previous system icons with Ionicons, providing a wider range of icons and better compatibility with the project. This change improves the visual appearance and consistency of the Room component.

Related to recent user commits:
- refactor: Remove console.log statement in createMessage action of roomSlice.ts
- feat: Update Room component to use react-native-get-random-values
- refactor: Use Clipboard to copy message text in Room component
- refactor: Update npm dependencies for react-native-clipboard to version 1.14.1
- refactor: Update Room component to handle delete action in ContextMenu
- refactor: Update Room component onPress event handling
- refactor: Update Room component to use themed ContextMenu
- feat: Add ContextMenu to Room component for message options
- refactor: Update dependencies for react-native-context-menu-view to version 1.16.0