refactor: Include redirect functionality in NotFoundScreen component
This commit adds a redirect functionality to the NotFoundScreen component. When the screen is rendered, a timer is started using the useEffect hook. After 5 seconds, the router.back() function is called to navigate back to the previous screen. This improvement enhances the user experience by automatically redirecting them after a certain period of time.

Related to recent user commits:
- refactor: Update TabsLayout component to include dynamic height for tabs
- fix: tabs style