refactor: Update authService to include resetPassword function
This commit updates the authService module to include a new function called resetPassword. This function allows users to reset their password by sending a recovery email to their registered email address. The function makes use of the createRecovery method from the account service in the apiService module. The email parameter is passed to the createRecovery method along with the NEW_PASSWORD_URL constant from the config file. If an error occurs during the password reset process, an error is thrown. This update improves the functionality of the authentication system by providing a way for users to reset their password.