Added Nativewind
It's really important to note, that the state the app is currently in is very buggy and after all pretty bad and not everything works good with nativewind right now, for example; the ThemedText already has some predefined styles, so if you try to use something like text-[20px] on it, it won't work, so we'll have to figure that out (it's not that hard, I've used nativewind before). Also I've been working with nativewind on one of my personal projects and have noticed, that you can't use things like m-1, but you have to do m-[4px] and it takes some aditional configuration to get the mentioned m-1 to work.